SDIOS/models_server/requirements.txt
# server
flask
werkzeug

# client
requests
tensorflow # convert h5 to tflite on upload

# keypair
pycryptodome
